Make stbt.FrameDiffer and stbt.MotionDiff public APIs
And move the implementations into `_stbt/diff.py` where additional
implementations will be able to share code.
