+++ Makefile
@@ -70,6 +70,7 @@ INSTALL_PYLIB_FILES = \
     _stbt/control.py \
     _stbt/core.py \
     _stbt/cv2_compat.py \
+    _stbt/diff.py \
     _stbt/frameobject.py \
     _stbt/grid.py \
     _stbt/gst_hacks.py \
